Fixed unsigned -> signed warnings
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@43604 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/arch/gtk3/widgets/videopalettewidget.c
@@ -177,7 +177,7 @@ static GtkWidget *create_combo_box(void)
         ext = util_get_extension(current);
         if ((ext != NULL) && (strcmp(ext, "vpl"))) {
             /* current already has .vpl extension, so remove it for the extra string */
-            int n = strlen(current);
+            size_t n = strlen(current);
             current2 = lib_strdup(current);
             current2[n - 4] = 0;
         } else if (ext == NULL) {
